withCollectionTracker -> CollectionTracker
